Tooltip position can be set on a per-step basis by setting an element's data-position to 'top', 'right', 'bottom', or 'left.
The "arrow" was refactored into its own element for cross-browser compatibility since :before can be problematic in <= IE8. Removed the timeout for moving the tooltip, as the delay was causing the tooltip not to size properly. Added _placeTooltip method that uses targetElement, tooltipLayer, and arrowLayer as parameters, then sets the appropriate CSS styles for the tooltipLayer.