Added read me instructions. Fixed default position.
Added instructions on how to use the 'data-position' attribute to the read me. Fixed the default position of the tooltip where it wasn't placing the tooltip at the bottom position unless specified as such; it now defaults to the bottom if no data-position is set. Ran the file through javascriptlint and fixed a couple small things.