Targets can now be highlighted even inside absolutely positioned elements with a low z-index.
Refactored the way highlighting works. Previously, the highlight was merely a CSS class with a high z-index that was applied to the targetElement. Since children cannot have higher z-indexes than their parents, this made it impossible to highlight certain elements

This patch replaces the highlight class with a more complex (and, frankly, a little tacky) method wherein the target element is copied, along with its styles and children, and placed into the helperLayer itself. This will likely cause some unanticipated issues, but these issues are something that can be resolved whereas the z-index issue with the highlight class could not have been.