Separate content + HTML
I have made use of jQuery as I am using it in my application, but I'm sure you can do it with pure JavaScript.

I have managed to insert HTML in the tour popups as well as completely separate the tour content from the page layout.
data-intro attributes are no longer needed. Instead, a list containing information about the elements to take part in the tour and its text is needed as well as the floating box position.

data-step is not either needed. It will follow the list elements order.