added onafterchange method, and call it just before the end of _showElement method
difference with onchange is that onchange is executed first time upon entering _showElement method. This will fail if we want to, for example, set focus to the current input element being shown, because the element is modified in _showElement method. onafterchange is executed just before the end of _showElement method, so any changes to the current element will persist. I found this especially useful when trying to set focus to an input element which is currently being intro-ed using jquery's .focus().