Fix issue 356 - no interaction with the target element.
Up until now (since it was added in pull request 273), the reference
layer had a higher z-index than the target element, which caused it to
block interaction with target element, as if disableInteraction is
enabled.

To solve this, the reference layer is now hidden, but its children are
made visible. This restores the ability to interact with the underlying
element.

Changing the visibility attribute doesn't affect positioning, so no
other changes were required.