Fix issue 369: step number position with floating tooltip
The 'floating' position type was added in be96db029, where it can only
be used when there's no target element. Since there's no visible element
on which to put the number, code was added to position the number layer
next to the tooltip layer in case of 'floating' position.

Since pull requst #273 however, the 'floating' position can also be used
*with* a target element if 'auto' positioning is used (i.e.
_determineAutoPosition can return 'floating'). In this case the code to
position the step number next to the tooltip no longer makes sense, and
it moves the number layer away from the target element.

This commit only positions the step number if there's no target element
(i.e. if the 'introjsFloatingElement' element is used), as initially
intended. If there is a target element, the step number is already
positioned correctly.

Also changed the step number positiion offset from 18 to 16 px to match
the css file.