fix(element): Skip an element if it is null.
The expectation is that a null element represents an
element that could not be selected, so the tour step should
be skipped.  An undefined element assumes that no element
was intended to be associated with the step, so a floating
modal should be used instead.  This matches the way the
return value for an element property with a function value
is treated.