Makes introjs work with transformed elements
Prior to this change, any time parent elements introduced a stacking
context with certain positioning or with transforms, introjs would just
forcibly remove those. For something like opacity, that may not be a big
deal, but when transforms are an integral part of page styling, you
can't just disable them and expect things to render properly.

In those cases, now we clone the highlighted element and plop it right
on top of the highlighting layer--sidestepping the whole z-index problem
to begin with.

This also yanks out the alternate method for positioning, since that
also did not take transforms into account. Now the highlight will be
positioned properly on the page, even when parents are scaled or
translated.