Fix offset of element if html element has overflow properties
When the <html> element has an overflow set (overflow: auto or scroll),
the _scrollParentToElement function is called.
This results in the element being at the top of the page.

This PR checks against the `html` tag as well as the `body` tag when
choosing which scroll method to use.

The result is that the element is nicely positioned in the center of the
screen.