Processing possible absence of an element by selector
When creating list of steps, we take into account the possible absence of element by selector

If .element is a selector for which there is no element at the configuration stage, then we replace .element specifying a getter/setter for it. In the getter  we return the element via document.querySelector

Since that moment we assume that .element should return HTMLElement by selector and if there is still not element - throwing Error. Thus to check raw field below we need use ._element